+++ ios/Classes/FlutterAppauthPlugin.m
@@ -21,6 +21,8 @@
 @property(nonatomic, strong) NSString *discoveryUrl;
 @property(nonatomic, strong) NSString *redirectUrl;
 @property(nonatomic, strong) NSString *refreshToken;
+@property(nonatomic, strong) NSString *codeVerifier;
+@property(nonatomic, strong) NSString *authorizationCode;
 @property(nonatomic, strong) NSArray *scopes;
 @property(nonatomic, strong) NSDictionary *serviceConfigurationParameters;
 @property(nonatomic, strong) NSDictionary *additionalParameters;
@@ -35,6 +37,8 @@
     _discoveryUrl = [ArgumentProcessor processArgumentValue:arguments withKey:@"discoveryUrl"];
     _redirectUrl = [ArgumentProcessor processArgumentValue:arguments withKey:@"redirectUrl"];
     _refreshToken = [ArgumentProcessor processArgumentValue:arguments withKey:@"refreshToken"];
+    _authorizationCode = [ArgumentProcessor processArgumentValue:arguments withKey:@"authorizationCode"];
+    _codeVerifier = [ArgumentProcessor processArgumentValue:arguments withKey:@"codeVerifier"];
     _grantType = [ArgumentProcessor processArgumentValue:arguments withKey:@"grantType"];
     _scopes = [ArgumentProcessor processArgumentValue:arguments withKey:@"scopes"];
     _serviceConfigurationParameters = [ArgumentProcessor processArgumentValue:arguments withKey:@"serviceConfiguration"];
@@ -63,13 +67,16 @@
 @implementation FlutterAppauthPlugin
 
 FlutterMethodChannel* channel;
-NSString *const AUTHORIZE_AND_EXCHANGE_TOKEN_METHOD = @"authorizeAndExchangeToken";
+NSString *const AUTHORIZE_METHOD = @"authorize";
+NSString *const AUTHORIZE_AND_EXCHANGE_CODE_METHOD = @"authorizeAndExchangeCode";
 NSString *const TOKEN_METHOD = @"token";
+NSString *const AUTHORIZE_ERROR_CODE = @"authorize_failed";
+NSString *const AUTHORIZE_AND_EXCHANGE_CODE_ERROR_CODE = @"authorize_and_exchange_code_failed";
 NSString *const DISCOVERY_ERROR_CODE = @"discovery_failed";
 NSString *const TOKEN_ERROR_CODE = @"token_failed";
 NSString *const DISCOVERY_ERROR_MESSAGE_FORMAT = @"Error retrieving discovery document: %@";
-NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
-
+NSString *const TOKEN_ERROR_MESSAGE_FORMAT = @"Failed to get token %@";
+NSString *const AUTHORIZE_ERROR_MESSAGE = @"Failed to authorize";
 
 
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
@@ -82,8 +89,10 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
 }
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
-    if ([AUTHORIZE_AND_EXCHANGE_TOKEN_METHOD isEqualToString:call.method]) {
-        [self handleAuthorizeMethodCall:[call arguments] result:result];
+    if ([AUTHORIZE_AND_EXCHANGE_CODE_METHOD isEqualToString:call.method]) {
+        [self handleAuthorizeMethodCall:[call arguments] result:result exchangeCode:true];
+    } else if([AUTHORIZE_METHOD isEqualToString:call.method]) {
+        [self handleAuthorizeMethodCall:[call arguments] result:result exchangeCode:false];
     } else if([TOKEN_METHOD isEqualToString:call.method]) {
         [self handleTokenMethodCall:[call arguments] result:result];
     } else {
@@ -91,14 +100,14 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
     }
 }
 
--(void)handleAuthorizeMethodCall:(NSDictionary*)arguments result:(FlutterResult)result {
+-(void)handleAuthorizeMethodCall:(NSDictionary*)arguments result:(FlutterResult)result exchangeCode:(BOOL)exchangeCode {
     AuthorizationTokenRequestParameters *requestParameters = [[AuthorizationTokenRequestParameters alloc] initWithArguments:arguments];
     if(requestParameters.serviceConfigurationParameters != nil) {
         OIDServiceConfiguration *serviceConfiguration =
         [[OIDServiceConfiguration alloc]
          initWithAuthorizationEndpoint:[NSURL URLWithString:requestParameters.serviceConfigurationParameters[@"authorizationEndpoint"]]
          tokenEndpoint:[NSURL URLWithString:requestParameters.serviceConfigurationParameters[@"tokenEndpoint"]]];
-        [self performAuthorization:serviceConfiguration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result];
+        [self performAuthorization:serviceConfiguration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result exchangeCode:exchangeCode];
     } else if (requestParameters.discoveryUrl) {
         NSURL *discoveryUrl = [NSURL URLWithString:requestParameters.discoveryUrl];
         
@@ -111,7 +120,7 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
                                                                           return;
                                                                       }
                                                                       
-                                                                      [self performAuthorization:configuration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result];
+                                                                      [self performAuthorization:configuration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result exchangeCode:exchangeCode];
                                                                   }];
     } else {
         NSURL *issuerUrl = [NSURL URLWithString:requestParameters.issuer];
@@ -124,14 +133,14 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
                                                                     return;
                                                                 }
                                                                 
-                                                                [self performAuthorization:configuration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result];
+                                                                [self performAuthorization:configuration clientId:requestParameters.clientId clientSecret:requestParameters.clientSecret scopes:requestParameters.scopes redirectUrl:requestParameters.redirectUrl additionalParameters:requestParameters.additionalParameters result:result exchangeCode:exchangeCode];
                                                             }];
     }
     
     
 }
 
-- (void)performAuthorization:(OIDServiceConfiguration *)serviceConfiguration clientId:(NSString*)clientId clientSecret:(NSString*)clientSecret scopes:(NSArray *)scopes redirectUrl:(NSString*)redirectUrl additionalParameters:(NSDictionary *)additionalParameters result:(FlutterResult)result {
+- (void)performAuthorization:(OIDServiceConfiguration *)serviceConfiguration clientId:(NSString*)clientId clientSecret:(NSString*)clientSecret scopes:(NSArray *)scopes redirectUrl:(NSString*)redirectUrl additionalParameters:(NSDictionary *)additionalParameters result:(FlutterResult)result exchangeCode:(BOOL)exchangeCode{
     OIDAuthorizationRequest *request =
     [[OIDAuthorizationRequest alloc] initWithConfiguration:serviceConfiguration
                                                   clientId:clientId
@@ -142,22 +151,42 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
                                       additionalParameters:additionalParameters];
     UIViewController *rootViewController =
     [UIApplication sharedApplication].delegate.window.rootViewController;
-    _currentAuthorizationFlow = [OIDAuthState authStateByPresentingAuthorizationRequest:request
-                                                               presentingViewController: rootViewController
-                                                                               callback:^(OIDAuthState *_Nullable authState,
-                                                                                          NSError *_Nullable error) {
-                                                                                   if(authState) {
-                                                                                       result([self processResponses:authState.lastTokenResponse authResponse:authState.lastAuthorizationResponse]);
-                                                                                   } else {
-                                                                                       
-                                                                                   }
-                                                                               }];
+    if(exchangeCode) {
+        _currentAuthorizationFlow = [OIDAuthState authStateByPresentingAuthorizationRequest:request presentingViewController: rootViewController
+                                                                                   callback:^(OIDAuthState *_Nullable authState,
+                                                                                              NSError *_Nullable error) {
+                                                                                       if(authState) {
+                                                                                           result([self processResponses:authState.lastTokenResponse authResponse:authState.lastAuthorizationResponse]);
+                                                                                           
+                                                                                       } else {
+                                                                                           [self finishWithError:AUTHORIZE_AND_EXCHANGE_CODE_ERROR_CODE message:AUTHORIZE_ERROR_MESSAGE result:result];
+                                                                                       }
+                                                                                   }];
+    } else {
+        _currentAuthorizationFlow = [OIDAuthorizationService presentAuthorizationRequest:request presentingViewController:rootViewController callback:^(OIDAuthorizationResponse *_Nullable authorizationResponse, NSError *_Nullable error) {
+            if(authorizationResponse) {
+                NSMutableDictionary *processedResponse = [[NSMutableDictionary alloc] init];
+                [processedResponse setObject:authorizationResponse.additionalParameters forKey:@"authorizationAdditionalParameters"];
+                [processedResponse setObject:authorizationResponse.authorizationCode forKey:@"authorizationCode"];
+                [processedResponse setObject:authorizationResponse.request.codeVerifier forKey:@"codeVerifier"];
+                result(processedResponse);
+            } else {
+                [self finishWithError:AUTHORIZE_ERROR_CODE message:AUTHORIZE_ERROR_MESSAGE result:result];
+            }
+        }];
+        
+    }
 }
 
 - (void)finishWithDiscoveryError:(NSError * _Nullable)error result:(FlutterResult)result {
     NSString *message = [NSString stringWithFormat:DISCOVERY_ERROR_MESSAGE_FORMAT, [error localizedDescription]];
-    result([FlutterError errorWithCode:DISCOVERY_ERROR_CODE message:message details:nil]);
+    [self finishWithError:DISCOVERY_ERROR_CODE message:message result:result];
 }
+
+- (void)finishWithError:(NSString *)errorCode message:(NSString *)message  result:(FlutterResult)result {
+    result([FlutterError errorWithCode:errorCode message:message details:nil]);
+}
+
 
 -(void)handleTokenMethodCall:(NSDictionary*)arguments result:(FlutterResult)result {
     TokenRequestParameters *requestParameters = [[TokenRequestParameters alloc] initWithArguments:arguments];
@@ -202,20 +231,21 @@ NSString *const TOKEN_ERROR_MESSAGE = @"Failed to exchange token";
     OIDTokenRequest *tokenRequest =
     [[OIDTokenRequest alloc] initWithConfiguration:serviceConfiguration
                                          grantType:requestParameters.grantType
-                                 authorizationCode:nil
+                                 authorizationCode:requestParameters.authorizationCode
                                        redirectURL:[NSURL URLWithString:requestParameters.redirectUrl]
                                           clientID:requestParameters.clientId
                                       clientSecret:requestParameters.clientSecret
                                             scopes:requestParameters.scopes
                                       refreshToken:requestParameters.refreshToken
-                                      codeVerifier:nil
+                                      codeVerifier:requestParameters.codeVerifier
                               additionalParameters:requestParameters.additionalParameters];
     [OIDAuthorizationService performTokenRequest:tokenRequest
                                         callback:^(OIDTokenResponse *_Nullable response,
                                                    NSError *_Nullable error) {
                                             if (response) {
                                                 result([self processResponses:response authResponse:nil]);                                           } else {
-                                                    result(nil);
+                                                    NSString *message = [NSString stringWithFormat:TOKEN_ERROR_MESSAGE_FORMAT, [error localizedDescription]];
+                                                    [self finishWithError:TOKEN_ERROR_CODE message:message result:result];
                                                 }
                                         }];
 }